+++ include/paa/paa.h
@@ -11,7 +11,7 @@
 #include "globals.h"
 
 
-Value *piecewiseAggregate(Value const *values, size_t size, size_t series_length, size_t summarization_length,
-                          int num_threads);
+Value *piecewiseAggregate(Value const *values, unsigned int size, unsigned int series_length, unsigned int summarization_length,
+                          unsigned int num_threads);
 
 #endif //ISAX_PAA_H
